+++ tests/local/strpad.php
@@ -1,0 +1,11 @@
+<?php
+declare (strict_types = 1);
+namespace WebSharks\Core\Test;
+
+use WebSharks\Core\Classes\CoreFacades as c;
+
+require_once dirname(__FILE__, 2).'/includes/local.php';
+
+/* ------------------------------------------------------------------------------------------------------------------ */
+
+echo c::mbStrPad('123', 20, '0', STR_PAD_LEFT);